Struct darkirc::irc::client::Client

pub struct Client {

    pub server: Arc<IrcServer>,
    pub incoming: Subscription<Event>,
    pub addr: SocketAddr,
    pub last_sent: RwLock<Hash>,
    pub channels: RwLock<HashSet<String>>,
    pub penalty: AtomicUsize,
    pub registered: AtomicBool,
    pub reg_paused: AtomicBool,
    pub username: RwLock<String>,
    pub nickname: RwLock<String>,
    pub realname: RwLock<String>,
    pub caps: RwLock<HashMap<String, bool>>,
    pub seen: OnceCell<Tree>,

}

Stateful IRC client handler, used for each client connection

Fields
server: Arc<IrcServer>

Pointer to parent IrcServer

incoming: Subscription<Event>

Subscription for incoming events

addr: SocketAddr

Client socket addr

last_sent: RwLock<Hash>

ID of the last sent event

channels: RwLock<HashSet<String>>

Active (joined) channels for this client

penalty: AtomicUsize

Penalty counter, when limit is reached, disconnect client

registered: AtomicBool

Registration marker

reg_paused: AtomicBool

Registration pause marker

username: RwLock<String>

Client username

nickname: RwLock<String>

Client nickname

realname: RwLock<String>

Client realname

caps: RwLock<HashMap<String, bool>>

Client caps

seen: OnceCell<Tree>

Set of seen messages for the user TODO: It grows indefinitely, needs to be pruned.

Implementations

impl Client

pub async fn new( server: Arc<IrcServer>, incoming: Subscription<Event>, addr: SocketAddr ) -> Result<Self>

Instantiate a new Client.

pub async fn multiplex_connection<S>(&self, stream: S) -> Result<()>where S: AsyncRead + AsyncWrite + Unpin + Send + 'static,

This function handles a single IRC client. We listen to messages from the IRC client and relay them to the network, and we also get notified of incoming messages and relay them to the IRC client. The notifications come from events being inserted into the Event Graph.

async fn reply<W>(&self, writer: &mut W, reply: &ReplyType) -> Result<()>where W: AsyncWrite + Unpin,

Send a reply to the IRC client. Matches on the reply type.

async fn process_client_line<W>( &self, line: &str, writer: &mut W ) -> Result<Option<Event>>where W: AsyncWrite + Unpin,

Handle the incoming line given sent by the IRC client

pub async fn mark_seen(&self, event_id: &Hash) -> Result<()>

Atomically mark a message as seen for this client.

pub async fn is_seen(&self, event_id: &Hash) -> Result<bool>

Check if a message was already marked seen for this client.

impl Client

pub async fn handle_cmd_admin(&self, _args: &str) -> Result<Vec<ReplyType>>

ADMIN [<server>]

Asks the server for information about the administrator of the server.

pub async fn handle_cmd_cap(&self, args: &str) -> Result<Vec<ReplyType>>

CAP <args>

pub async fn handle_cmd_info(&self, _args: &str) -> Result<Vec<ReplyType>>

INFO [<target>]

Gives information about the <target> server, or the current server if <target> is not used. The information includes the server’s version, when it was compiled, the patch level, when it was started, and any other information which might be relevant.

pub async fn handle_cmd_join(&self, args: &str) -> Result<Vec<ReplyType>>

JOIN <channels> [<keys>]

Makes the client join the channels in the list <channels>. Passwords can be used in the list <keys>. If the channels do not exist, they will be created.

pub async fn handle_cmd_list(&self, _args: &str) -> Result<Vec<ReplyType>>

LIST [<channels> [<server>]]

List all channels on the server. If the list <channels> is given, it will return the channel topics. If is given, the command will be sent to <server> for evaluation.

pub async fn handle_cmd_mode(&self, args: &str) -> Result<Vec<ReplyType>>

MODE <nickname> <flags> MODE <channel> <flags>

The MODE command has two uses. It can be used to set both user and channel modes.

pub async fn handle_cmd_motd(&self, _args: &str) -> Result<Vec<ReplyType>>

MOTD [<server>]

Returns the message of the day on or the current server if it is not stated.

pub async fn handle_cmd_names(&self, args: &str) -> Result<Vec<ReplyType>>

NAMES [<channel>]

Returns a list of who is on the list of , by channel name. If is not used, all users are shown. They are grouped by channel name with all users who are not on a channel being shown as part of channel “*”.

pub async fn handle_cmd_nick(&self, args: &str) -> Result<Vec<ReplyType>>

NICK <nickname>

Allows a client to change their IRC nickname.

pub async fn handle_cmd_part(&self, args: &str) -> Result<Vec<ReplyType>>

PART <channel>

Causes a user to leave the channel .

pub async fn handle_cmd_ping(&self, args: &str) -> Result<Vec<ReplyType>>

PING <server1>

Tests a connection. A PING message results in a PONG reply.

pub async fn handle_cmd_privmsg(&self, args: &str) -> Result<Vec<ReplyType>>

PRIVMSG <msgtarget> <message>

Sends to . The target is usually a user or a channel.

pub async fn handle_cmd_rehash(&self, _args: &str) -> Result<Vec<ReplyType>>

REHASH

Causes the server to re-read and re-process its configuration file(s).

pub async fn handle_cmd_topic(&self, args: &str) -> Result<Vec<ReplyType>>

TOPIC <channel> [<topic>]

Used to get the channel topic on . If is given, it sets the channel topic to .

pub async fn handle_cmd_user(&self, args: &str) -> Result<Vec<ReplyType>>

USER <user> <mode> <unused> <realname>

This command is used at the beginning of a connection to specify the username, hostname, real name, and the initial user modes of the connecting client. may contain spaces, and thus must be prefixed with a colon.

pub async fn handle_cmd_version(&self, _args: &str) -> Result<Vec<ReplyType>>

VERSION

Returns the version of the server.

async fn welcome(&self) -> Vec<ReplyType>

Internal function that constructs the welcome message.

async fn get_history( &self, channels: &HashSet<String> ) -> Result<Vec<ReplyType>>

Internal function that scans the DAG and returns events for given channels. Will return empty if no_history CAP is requested.