Module darkirc::irc::command

IRC command handler IRC command implemenatations

These try to follow the RFCs, modified in order for our P2P stack. Copied from https://simple.wikipedia.org/wiki/List_of_Internet_Relay_Chat_commands

Unimplemented commands:

AWAY

CONNECT

DIE

ERROR

INVITE

ISON

KICK

KILL

NOTICE

OPER

PASS

RESTART

SERVICE

SERVLIST

SERVER

SQUERY

SQUIT

SUMMON

TRACE

USERHOST

WALLOPS

WHO

WHOIS

WHOWAS

Some of the above commands could actually be implemented and could work in respect to the P2P network.