Module darkirc::irc::rpl

IRC numerics and server replies

Constants

ALREADY_REGISTERED

The message sent to the client when they are already registered

ERR_ALREADYREGISTERED

<client> :You may not reregister

ERR_ERRONEOUSNICKNAME

:Erroneus nickname

ERR_NEEDMOREPARAMS

<client> <command> :Not enough parameters

ERR_NOORIGIN

<client> :No origin specified

ERR_NORECIPIENT

<client> :No recipient given (<command>)

ERR_NOSUCHCHANNEL

<client> <channel> :No such channel

ERR_NOSUCHNICK

<client> <nickname> :No such nick/channel

ERR_NOTEXTTOSEND

<client> :No text to send

ERR_NOTREGISTERED

<client> :You have not registered

ERR_USERSDONTMATCH

<client> :Cant change mode for other users

INVALID_SYNTAX

The message sent to the client when command params could not parse

NOT_REGISTERED

The message sent to the client when they are not registered

RPL_ADMINEMAIL

<client> :<info>

RPL_ADMINLOC1

<client> :<info>

RPL_ADMINLOC2

<client> :<info>

RPL_ADMINME

<client> [<server>] :Administrative info

RPL_CHANNELMODEIS

<client> <channel> <modestring> <mode arguments>...

RPL_ENDOFINFO

<client> :End of INFO list

RPL_ENDOFMOTD

<client> :End of /MOTD command.

RPL_ENDOFNAMES

<client> <channel> :End of /NAMES list

RPL_INFO

<client> :<string>

RPL_LIST

<client> <channel> <client count> :<topic>

RPL_LISTEND

:End of /LIST

RPL_LISTSTART

<client> Channel :Users Name

RPL_MOTD

<client> :<line of the motd>

RPL_MOTDSTART

<client> :- <server> Message of the day -

RPL_NAMREPLY

<client> <symbol> <channel> :[prefix]<nick>{ [prefix]<nick>}

RPL_NOTOPIC

<client> <channel> :No topic is set

RPL_TOPIC

<client> <channel> :<topic>

RPL_UMODEIS

<client> <user modes>

RPL_VERSION

<client> <version> <server> :<comments>

RPL_WELCOME

<client> :Welcome to the DarkIRC network

RPL_YOURHOST

<client> :Your host is <servername>, running version <version

WELCOME

The welcome message sent upon successful registration